Refactor
* Make CLI separate module
* Defer imports for CLI
* Make all modules private that are not part of the public API
