+++ pyproject.toml
@@ -67,7 +67,6 @@ urls.Documentation = "https://github.com/Boehringer-Ingelheim/dso"
 urls.Home-page = "https://github.com/Boehringer-Ingelheim/dso"
 urls.Source = "https://github.com/Boehringer-Ingelheim/dso"
 scripts.dso = "dso.cli:cli"
-scripts.dso_pandocfilter = "dso.pandocfilter:main"
 
 [tool.hatch.version]
 source = "vcs"
