+++ tests/test_exec.py
@@ -4,7 +4,7 @@ from os import chdir
 import pytest
 from click.testing import CliRunner
 
-from dso.exec import cli
+from dso.exec import exec_cli
 
 
 @pytest.mark.parametrize("quiet", [None, "2"])
@@ -21,7 +21,7 @@ def test_exec_quarto(quarto_stage, quiet, launch_dir):
     if quiet is not None:
         os.environ["DSO_QUIET"] = quiet
 
-    result = runner.invoke(cli, ["quarto", stage_path])
+    result = runner.invoke(exec_cli, ["quarto", stage_path])
     assert result.exit_code == 0
     assert (quarto_stage / "report" / "quarto_stage.html").is_file()
     assert "Hello World!" in (quarto_stage / "report" / "quarto_stage.html").read_text()
@@ -32,7 +32,7 @@ def test_exec_quarto_empty_params(quarto_stage_empty_configs):
     chdir(quarto_stage_empty_configs)
     stage_path = "."
 
-    result = runner.invoke(cli, ["quarto", stage_path])
+    result = runner.invoke(exec_cli, ["quarto", stage_path])
     assert result.exit_code == 0
     assert (quarto_stage_empty_configs / "report" / "quarto_stage.html").is_file()
     assert "Hello World!" in (quarto_stage_empty_configs / "report" / "quarto_stage.html").read_text()
@@ -43,7 +43,7 @@ def test_exec_quarto_bibliography(quarto_stage_bibliography):
     chdir(quarto_stage_bibliography)
     stage_path = "."
 
-    result = runner.invoke(cli, ["quarto", stage_path])
+    result = runner.invoke(exec_cli, ["quarto", stage_path])
     assert result.exit_code == 0
     assert (quarto_stage_bibliography / "report" / "quarto_stage.html").is_file()
     assert "Knuth" in (quarto_stage_bibliography / "report" / "quarto_stage.html").read_text()
@@ -54,7 +54,7 @@ def test_exec_quarto_stylesheet(quarto_stage_css):
     chdir(quarto_stage_css)
     stage_path = "."
 
-    result = runner.invoke(cli, ["quarto", stage_path])
+    result = runner.invoke(exec_cli, ["quarto", stage_path])
     assert result.exit_code == 0
     assert (quarto_stage_css / "report" / "quarto_stage.html").is_file()
     assert "h2.veryspecialclass1" in (quarto_stage_css / "report" / "quarto_stage.html").read_text()
